mptcp: fix 'masking a bool' warning
Git-commit: c4512c63b1193c73b3f09c598a6d0a7f88da1dd8
Message-Id: <c4512c63b1193c73b3f09c598a6d0a7f88da1dd8.1637072333.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.14-rc1

Dan Carpenter reported an issue introduced in
commit fde56eea01f9 ("mptcp: refine mptcp_cleanup_rbuf") where a new
boolean (ack_pending) is masked with 0x9.

This is not the intention to ignore values by using a boolean. This
variable should not have a 'bool' type: we should keep the 'u8' to allow
this comparison.

Fixes: fde56eea01f9 ("mptcp: refine mptcp_cleanup_rbuf")
Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Acked-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.14-rc1-mptcp-fix-masking-a-bool-warning.patch